Final Commit, Fixed Unauthorized Access To Pages and Movie Operations, Added Cron Job To Delete Inactive Users Ever 12Hrs

+++ manflix/routes.py
@@ -1,9 +1,8 @@
 from flask import Flask, render_template, request, redirect, flash, url_for, session
 from flask_login import login_user, current_user, logout_user, login_required
-import random
-import smtplib
-import secrets
-import os
+from apscheduler.schedulers.background import BackgroundScheduler
+from random import randint
+from smtplib import SMTP
 from manflix import app, db, bcrypt
 from manflix.functions import get_movie_detail, get_movie_id
 from manflix.forms import AddMovieForm, SearchMovieForm, RegistrationForm, LoginForm, VerifyUserForm, AdminForm, UpdateAccountForm, UpdateMovieForm
@@ -19,7 +18,7 @@ def index():
     else:
         return render_template('index.html',all_movies=all_movies)
 
-# Login
+# Login Page
 @app.route('/login',methods = ['GET','POST'])
 def login():
     if current_user.is_authenticated:
@@ -36,7 +35,7 @@ def login():
             flash(f'Please Check For Username and Password', 'danger')        
     return render_template('login.html',title='Login',form=form)
 
-# User Registration Form
+# User Registration Page
 @app.route('/register',methods = ['GET','POST'])
 def register():
     if current_user.is_authenticated:
@@ -46,17 +45,17 @@ def register():
         hashed_password = bcrypt.generate_password_hash(form.password.data).decode('utf-8') 
         user_name = form.username.data  
         e_mail = form.email.data   
-        otp = ''.join([str(random.randint(0, 9)) for i in range(4)])
-        user_avatar = form.user_avatar.data + '.webp'
-        print(user_avatar)
-        print(type(user_avatar))        
+
+        # Generating OTP
+        otp = ''.join([str(randint(0, 9)) for i in range(4)])
+        user_avatar = form.user_avatar.data + '.webp'   
         user = UserData(username=user_name,email=e_mail,password=hashed_password,otp=otp,avatar=user_avatar)
 
         # Adding Current Users Email To Session
         session["email"] = form.email.data
 
         # Sending OTP To Mail 
-        smtp_server = smtplib.SMTP('smtp.gmail.com', 587)
+        smtp_server = SMTP('smtp.gmail.com', 587)
         smtp_server.ehlo()
         smtp_server.starttls()
         smtp_server.login('johnharrison12587@gmail.com', 'fhwrnydeedhztaso')
@@ -135,14 +134,18 @@ def logout():
 @login_required
 def home():
     page = request.args.get('page',1,type = int)
-    action_movies = Movies.query.filter(Movies.category.contains('Action')).paginate(per_page = 3)
-    adventure_movies = Movies.query.filter(Movies.category.contains('Adventure')).paginate(per_page = 3)
-    animation_movies = Movies.query.filter(Movies.category.contains('Animation')).paginate(per_page = 3)
-    science_movies = Movies.query.filter(Movies.category.contains('Science')).paginate(per_page = 3)
-    return render_template('home.html',act_mov=action_movies,
+    action_movies = Movies.query.filter(Movies.category.contains('Action')).limit(5)
+    adventure_movies = Movies.query.filter(Movies.category.contains('Adventure')).limit(5)
+    animation_movies = Movies.query.filter(Movies.category.contains('Animation')).limit(5)
+    science_movies = Movies.query.filter(Movies.category.contains('Science')).limit(5)
+    recent_movies = Movies.query.order_by(Movies.id.desc()).limit(5)
+
+    return render_template('home.html',
+    act_mov=action_movies,
     adv_mov = adventure_movies,
     ani_mov = animation_movies,
-    sci_mov = science_movies
+    sci_mov = science_movies,
+    rect_mov = recent_movies 
     )
 
 # Route For Categories Of Movies
@@ -152,20 +155,6 @@ def category(cat):
     m = cat
     movie = Movies.query.filter(Movies.category.contains(cat)).paginate(per_page = 1)
     return render_template('category_movie.html',movie=movie,m=m)
-
-# Like Movie Function
-@app.route('/like/<int:mid>/<action>')
-@login_required
-def like_action(mid, action):
-    movie = Movies.query.filter_by(id=mid).first_or_404()
-    if action == 'like':
-        current_user.like_movie(movie)
-        flash(f'{movie.title} - Added To Favourites','success')
-        db.session.commit()
-    if action == 'unlike':
-        current_user.unlike_movies(movie)
-        db.session.commit()
-    return redirect(request.referrer)
 
 # Liked Movie Page
 @app.route('/liked_movies')
@@ -179,41 +168,45 @@ def liked_movies():
 @app.route('/add_movies', methods=['GET', 'POST'])
 @login_required
 def add_movies():
-    form1 = AddMovieForm()    
-    if request.method == 'POST':
-        if form1.validate_on_submit():
-            link = form1.link.data
-            quality = form1.quality.data
-            movie_id_tmd = form1.movie_id_tmd.data
-            dolby_audio = form1.dolby_audio.data
-            dual_audio = form1.dual_audio.data
-            poster_link = get_movie_detail(1, movie_id_tmd)
-            title = get_movie_detail(2, movie_id_tmd)            
-            movie_release_year = get_movie_detail(3, movie_id_tmd)
-            movie_backdrop_link = get_movie_detail(4, movie_id_tmd)            
-            trailer_link = get_movie_detail(6, movie_id_tmd)
-            category = get_movie_detail(7, movie_id_tmd)
-            movie = Movies(title=title, link=link,category=category, quality=quality,movie_release_year=movie_release_year, movie_id_tmd=movie_id_tmd, poster_link=poster_link, movie_backdrop_link=movie_backdrop_link,trailer_link=trailer_link, dolby_audio=dolby_audio, dual_audio=dual_audio)
-            db.session.add(movie)
-            db.session.commit()
-            flash(f'{ title } - Added Successfully!', 'success')
-            return render_template('add_movies.html', form1=form1)   
-        else:
-            search = request.form["searched"]    
-            movie_id = get_movie_id(search)    
-            if movie_id:
-                return render_template('add_movies.html', movie_id=movie_id, search=search,form1=form1)
+    if current_user.is_admin == True:
+        form1 = AddMovieForm()    
+        if request.method == 'POST':
+            if form1.validate_on_submit():
+                link = form1.link.data
+                quality = form1.quality.data
+                movie_id_tmd = form1.movie_id_tmd.data
+                dolby_audio = form1.dolby_audio.data
+                dual_audio = form1.dual_audio.data
+                poster_link = get_movie_detail(1, movie_id_tmd)
+                title = get_movie_detail(2, movie_id_tmd)            
+                movie_release_year = get_movie_detail(3, movie_id_tmd)
+                movie_backdrop_link = get_movie_detail(4, movie_id_tmd)            
+                trailer_link = get_movie_detail(6, movie_id_tmd)
+                category = get_movie_detail(7, movie_id_tmd)
+                movie = Movies(title=title, link=link,category=category, quality=quality,movie_release_year=movie_release_year, movie_id_tmd=movie_id_tmd, poster_link=poster_link, movie_backdrop_link=movie_backdrop_link,trailer_link=trailer_link, dolby_audio=dolby_audio, dual_audio=dual_audio)
+                db.session.add(movie)
+                db.session.commit()
+                flash(f'{ title } - Added Successfully!', 'success')
+                return render_template('add_movies.html', form1=form1)   
             else:
-                flash(f'{ search } - Not Found!', 'danger')
-                return render_template('add_movies.html', form1=form1)                  
-    return render_template('add_movies.html', form1=form1)
+                search = request.form["searched"]    
+                movie_id = get_movie_id(search)    
+                if movie_id:
+                    return render_template('add_movies.html', movie_id=movie_id, search=search,form1=form1)
+                else:
+                    flash(f'{ search } - Not Found!', 'danger')
+                    return render_template('add_movies.html', form1=form1)                  
+    return redirect(url_for('home'))
 
 # Movie Database Page
 @app.route('/movie_database')
 @login_required
-def movie_database():    
-    all_movies = Movies.query.all()
-    return render_template('movie_database.html', all_movies=all_movies)
+def movie_database():  
+    if current_user.is_admin == True:            
+        all_movies = Movies.query.all()
+        return render_template('movie_database.html', all_movies=all_movies)
+    else:
+        return redirect(url_for('home'))
 
 # Search Movie Page In Database
 @app.route('/search_movies', methods=['POST'])
@@ -241,35 +234,67 @@ def movie_screen(id):
 
 #------------------- All Functions ---------------------
 
+# Delete Inactive User
+def del_inactive_user():
+    UserData.query.filter_by(is_verified=False).delete() 
+    db.session.commit()
+    print('Deleted InActive User!')
+
+sched = BackgroundScheduler(daemon=True)
+sched.add_job(del_inactive_user,'interval',minutes=1440)
+sched.start()
+
+# Like Movie Function
+@app.route('/like/<int:mid>/<action>')
+@login_required
+def like_action(mid, action):
+    movie = Movies.query.filter_by(id=mid).first_or_404()
+    if action == 'like':
+        current_user.like_movie(movie)
+        flash(f'{movie.title} - Added To Favourites','success')
+        db.session.commit()
+    if action == 'unlike':
+        current_user.unlike_movies(movie)
+        db.session.commit()
+    return redirect(request.referrer)
+
 # Delete Movie Function
 @app.route('/delete/<int:id>')
 @login_required
 def delete(id):
-    movie = Movies.query.filter_by(id=id).first()
-    db.session.delete(movie)
-    db.session.commit()
-    return render_template('movie_database.html')
+    if current_user.is_admin == True:
+        movie = Movies.query.filter_by(id=id).first()
+        db.session.delete(movie)
+        db.session.commit()
+    else:
+        flash('You Are not Admin!','info')
+        return redirect(url_for('home'))
+    return redirect(url_for('movie_database'))
 
 # Update Movie Function
 @app.route('/update/<int:id>', methods=['GET', 'POST'])
 @login_required
 def update(id):
-    movie = Movies.query.filter_by(id=id).first()
-    form1 = UpdateMovieForm(obj=movie)    
-    if request.method == 'POST':
-        if form1.validate_on_submit():           
-            form1.populate_obj(movie)
-            movie_id_tmd = form1.movie_id_tmd.data
-            if movie_id_tmd != movie.movie_id_tmd:
-                print('Getting New Data')
-                movie.poster_link = get_movie_detail(1, movie_id_tmd)
-                movie.title = get_movie_detail(2, movie_id_tmd)
-                movie.movie_release_year = get_movie_detail(3, movie_id_tmd)
-                movie.movie_backdrop_link = get_movie_detail(4, movie_id_tmd)
-                movie.trailer_link = get_movie_detail(6, movie_id_tmd)
-            db.session.commit()
-            flash(f'Updated Successfully', 'success')  
-    return render_template('update_movie.html', movie=movie,form1=form1)
+    if current_user.is_admin == True:
+        movie = Movies.query.filter_by(id=id).first()
+        form1 = UpdateMovieForm(obj=movie)    
+        if request.method == 'POST':
+            if form1.validate_on_submit():           
+                form1.populate_obj(movie)
+                movie_id_tmd = form1.movie_id_tmd.data
+                if movie_id_tmd != movie.movie_id_tmd:
+                    print('Getting New Data')
+                    movie.poster_link = get_movie_detail(1, movie_id_tmd)
+                    movie.title = get_movie_detail(2, movie_id_tmd)
+                    movie.movie_release_year = get_movie_detail(3, movie_id_tmd)
+                    movie.movie_backdrop_link = get_movie_detail(4, movie_id_tmd)
+                    movie.trailer_link = get_movie_detail(6, movie_id_tmd)
+                db.session.commit()
+                flash(f'Updated Successfully', 'success')  
+        return render_template('update_movie.html', movie=movie,form1=form1)
+    else:
+        flash('You Are not Admin!','info')
+        return redirect(url_for('home'))
 
 
 #------------------- All Error Pages ---------------------
